feat: make field items of post data list extendable
